+++ shuup_tests/xtheme/test_views.py
@@ -7,6 +7,7 @@
 # LICENSE file in the root directory of this source tree.
 import pytest
 
+from shuup.testing import factories
 from shuup.utils.excs import Problem
 from shuup.xtheme.editing import is_edit_mode
 from shuup.xtheme.views.command import command_dispatch
@@ -16,6 +17,7 @@ from shuup_tests.utils.faux_users import SuperUser
 def test_edit_can_be_set_via_view(rf):
     request = rf.get("/")
     request.user = SuperUser()
+    request.shop = factories.get_default_shop()
     request.session = {}
     request.POST = {"command": "edit_on"}
     command_dispatch(request)