+++ doc/changelog.rst
@@ -12,6 +12,9 @@ Unrealeased
 Core
 ~~~~
 
+- Refund: check the max refundable items when doing partial refunds
+- Add customer related fields account manager, tax group and customer
+  groups to order for sales reporting.
 - Order source: consider the packages quantities in order source lines
 - Report: change date filter field to DateTime
 - OrderCreator: Dispatch a signal when adding lines to order
@@ -51,6 +54,7 @@ Core
 Admin
 ~~~~~
 
+- Improve admin order creator translations
 - Add option to filter variation parents for product select view
 - Fix home view help blocks filtering objects by the current shop
 - Fix bug: Filter category parent choices based on current shop
@@ -65,6 +69,7 @@ Admin
 Front
 ~~~~~
 
+- Add util for checking whether current user is admin
 - Limit reqular user login access to own specified shop only
 
 Campaigns